+++ src/main/java/com/example/dw/domain/dto/goods/GoodsDetailDto.java
@@ -10,8 +10,8 @@ public class GoodsDetailDto {
 
     private Long id;
     private String goodsName;
-    private Long goodsQuantity;
-    private Long goodsPrice;
+    private int goodsQuantity;
+    private int goodsPrice;
     private String goodsMade;
     private String goodsCertify;
     private String goodsDetailContent;
@@ -28,11 +28,8 @@ public class GoodsDetailDto {
     private String goodsDetailImgUuid;
 
     @QueryProjection
-    public GoodsDetailDto(Long id, String goodsName, Long goodsQuantity, Long goodsPrice,
-                          String goodsMade, String goodsCertify, String goodsDetailContent, String goodsRegisterDate,
-                          String goodsModifyDate, String goodsCategory, Long goodsMainImgId,String goodsMainImgName, String goodsMainImgPath,
-                          String goodsMainImgUuid, Long goodsDetailImgId, String goodsDetailImgName, String goodsDetailImgPath,
-                          String goodsDetailImgUuid) {
+    public GoodsDetailDto(Long id, String goodsName, int goodsQuantity, int goodsPrice, String goodsMade, String goodsCertify, String goodsDetailContent, String goodsRegisterDate, String goodsModifyDate, String goodsCategory, Long goodsMainImgId,
+                          String goodsMainImgName, String goodsMainImgPath, String goodsMainImgUuid, Long goodsDetailImgId, String goodsDetailImgName, String goodsDetailImgPath, String goodsDetailImgUuid) {
         this.id = id;
         this.goodsName = goodsName;
         this.goodsQuantity = goodsQuantity;
@@ -43,7 +40,7 @@ public class GoodsDetailDto {
         this.goodsRegisterDate = goodsRegisterDate;
         this.goodsModifyDate = goodsModifyDate;
         this.goodsCategory = goodsCategory;
-        this.goodsMainImgId= goodsMainImgId;
+        this.goodsMainImgId = goodsMainImgId;
         this.goodsMainImgName = goodsMainImgName;
         this.goodsMainImgPath = goodsMainImgPath;
         this.goodsMainImgUuid = goodsMainImgUuid;
